requirements.txt
Flask==3.0.0
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
Flask-Bootstrap4==4.0.2
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Login==0.6.3
Werkzeug==3.0.1
email_validator==2.1.0